+++ src/config/recommended.ts
@@ -16,6 +16,7 @@ module.exports = {
 		"empty-heading": "error",
 		"empty-title": "error",
 		"long-title": "error",
+		"meta-refresh": "error",
 		"no-conditional-comment": "error",
 		"no-deprecated-attr": "error",
 		"no-dup-attr": "error",